refactor: move set, modulegraph, validations, designtokens to internal/
These packages are implementation details, not public API. Moving them
under internal/ clarifies the public vs private API surface.

Closes #305
