+++ generate/classMemberDeclarations.go
@@ -25,7 +25,7 @@ import (
 	"bennypowers.dev/cem/generate/jsdoc"
 	M "bennypowers.dev/cem/manifest"
 	Q "bennypowers.dev/cem/queries"
-	S "bennypowers.dev/cem/set"
+	S "bennypowers.dev/cem/internal/set"
 
 	ts "github.com/tree-sitter/go-tree-sitter"
 )
