+++ generate/generate.go
@@ -24,7 +24,7 @@ import (
 	"slices"
 	"sync"
 
-	DT "bennypowers.dev/cem/designtokens"
+	DT "bennypowers.dev/cem/internal/designtokens"
 	DD "bennypowers.dev/cem/generate/demodiscovery"
 	M "bennypowers.dev/cem/manifest"
 	Q "bennypowers.dev/cem/queries"
