+++ internal/designtokens/testdata/draft-aliases/expected.json
