+++ internal/designtokens/testdata/draft-aliases/tokens.json
