+++ internal/designtokens/testdata/v2025-structured-colors/tokens.json
