+++ internal/modulegraph/module_graph_interfaces.go
