+++ internal/modulegraph/module_graph_lazy.go
