+++ internal/modulegraph/module_graph_metrics.go
