+++ internal/modulegraph/performance_test.go
