+++ internal/validations/data/global_attributes.json
