+++ internal/validations/global_attributes.go
