+++ internal/validations/global_attributes_test.go
