+++ lsp/types/context.go
@@ -19,7 +19,7 @@ package types
 import (
 	"bennypowers.dev/cem/internal/platform"
 	M "bennypowers.dev/cem/manifest"
-	"bennypowers.dev/cem/modulegraph"
+	"bennypowers.dev/cem/internal/modulegraph"
 	"bennypowers.dev/cem/queries"
 	protocol "github.com/tliron/glsp/protocol_3_16"
 )
